fixed process to manage images in StoreService

+++ app/Services/StoreService.php
@@ -89,15 +89,17 @@ class StoreService
 
         if ($request->hasAny($images)) {
             collect($images)->each(function ($file) use ($store, $request) {
-                $currentMedia =  $store->medias()->where('orientation', $file)->first();
-                if (!is_null($currentMedia)) {
-                    // TODO :: improvement process to get media
-                    $this->mediaService->delete($currentMedia->path);
+                if ($request->has($file)) {
+                    $currentMedia =  $store->medias()->where('orientation', $file)->first();
+                    if (!is_null($currentMedia)) {
+                        // TODO :: improvement process to get media
+                        $this->mediaService->delete($currentMedia->path);
+                    }
+                    $this->mediaService->save($request->file($file), [
+                        'orientation' => $file,
+                        'type_id' => Media::IMAGE,
+                    ]);
                 }
-                $this->mediaService->save($request->file($file), [
-                    'orientation' => $file,
-                    'type_id' => Media::IMAGE,
-                ]);
             });
 
             $store->load('medias');
